+++ pages/404.js
@@ -1,6 +1,6 @@
 import Link from 'next/link'
-import Layout from '../components/layout'
-import Button from '../components/button'
+import Layout from '../components/Layout'
+import Button from '../components/Button'
 import {
   errorHeading,
   errorSubHeadingTop,
