+++ pages/projects/[slug].js
@@ -3,13 +3,13 @@ import PropTypes from 'prop-types'
 import Head from 'next/head'
 import ErrorPage from 'next/error'
 import { getProjectBySlug, getAllProjects } from '../../lib/api'
-import Layout from '../../components/layout'
-import ProjectTitle from '../../components/project-title'
-import ProjectSubTitle from '../../components/project-subtitle'
-import ProjectParagraph from '../../components/project-paragraph'
-import ProjectBody from '../../components/project-body'
-import ProjectNav from '../../components/project-nav'
-import Divider from '../../components/divider'
+import Layout from '../../components/Layout'
+import ProjectTitle from '../../components/ProjectTitle'
+import ProjectSubTitle from '../../components/ProjectSubTitle'
+import ProjectParagraph from '../../components/ProjectParagraph'
+import ProjectBody from '../../components/ProjectBody'
+import ProjectNav from '../../components/ProjectNav'
+import Divider from '../../components/Divider'
 
 export default function Project({ project, projects }) {
   const {
