+++ pages/resume.js
@@ -1,9 +1,9 @@
 import Head from 'next/head'
-import Layout from '../components/layout'
-import ResumeList from '../components/resume-list'
-import ResumeHeading from '../components/resume-heading'
-import ResumeCopy from '../components/resume-copy'
-import Divider from '../components/divider'
+import Layout from '../components/Layout'
+import ResumeList from '../components/ResumeList'
+import ResumeHeading from '../components/ResumeHeading'
+import ResumeCopy from '../components/ResumeCopy'
+import Divider from '../components/Divider'
 import resume from '../data/resume'
 import { author, resumePage } from '../config/website'
